slettet endpoint saksjournal
- erstattet av: /sak/{saksnummer}/journal
